feature: add protocol serializable interface author: wunhwantseng@gmail.com
