+++ hawthorn-core/src/main/java/com/wunhwan/hawthorn/core/metadata/MethodMetadata.java
@@ -1,6 +1,9 @@
 package com.wunhwan.hawthorn.core.metadata;
 
+import com.wunhwan.hawthorn.core.annotation.RSocketService;
 import io.rsocket.frame.FrameType;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 import java.lang.reflect.Method;
 import java.util.Optional;
@@ -11,16 +14,53 @@ import java.util.Optional;
  * @author wunhwantseng@gmail.com
  * @since todo...
  */
-public interface MethodMetadata {
+public class MethodMetadata {
 
-    Method method();
+    private final Method method;
+    private final String endpoint;
+    private final int paramCount;
+    private final FrameType frameType;
+    private final Class<?> returnType;
 
-    String endpoint();
+    public MethodMetadata(Method method) {
+        this.method = method;
+        this.paramCount = method.getParameterCount();
+        this.returnType = method.getReturnType();
 
-    int paramCount();
+        if (method.isAnnotationPresent(RSocketService.class)) {
+            final RSocketService annotation = method.getAnnotation(RSocketService.class);
 
-    Optional<FrameType> frameType();
+            this.endpoint = annotation.endpoint();
+        } else {
+            this.endpoint = null;
+        }
 
-    Class<?> returnType();
+        if (Mono.class.equals(returnType)) {
+            this.frameType = FrameType.REQUEST_RESPONSE;
+        } else if (Flux.class.equals(returnType)) {
+            this.frameType = FrameType.REQUEST_CHANNEL;
+        } else {
+            this.frameType = FrameType.REQUEST_RESPONSE;
+        }
+    }
 
+    public Method method() {
+        return this.method;
+    }
+
+    public String endpoint() {
+        return this.endpoint;
+    }
+
+    public int paramCount() {
+        return this.paramCount;
+    }
+
+    public Optional<FrameType> frameType() {
+        return Optional.of(frameType);
+    }
+
+    public Class<?> returnType() {
+        return this.returnType;
+    }
 }
