+++ hawthorn-core/src/main/java/com/wunhwan/hawthorn/core/metadata/RoutingMetadata.java
@@ -1,0 +1,10 @@
+package com.wunhwan.hawthorn.core.metadata;
+
+/**
+ * todo...
+ *
+ * @author wunhwantseng@gmail.com
+ * @since todo...
+ */
+public class RoutingMetadata {
+}
